requirements.txt
fastapi
uvicorn
python-dotenv
llama-index
llama-parse
python-multipart
nest-asyncio